requirements.txt
streamlit
transformers
pandas
python-docx
fpdf
yake
PyPDF2